requirements-test.txt
# Test Framework Dependencies
# For comprehensive crypto data collection testing

# Core Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1        # Parallel test execution
pytest-timeout==2.1.0       # Test timeouts
pytest-mock==3.11.1         # Mocking utilities
pytest-html==4.1.1          # HTML test reports
pytest-benchmark==4.0.0     # Performance benchmarking

# Docker and Container Testing
docker==6.1.3               # Docker Python SDK
testcontainers==3.7.1       # Testcontainers for integration testing

# Database Testing
mysql-connector-python==8.2.0  # MySQL driver
PyMySQL==1.1.0              # Alternative MySQL driver
redis==5.0.1                # Redis client
SQLAlchemy==2.0.23          # Database ORM for complex queries

# HTTP and API Testing  
requests==2.31.0            # HTTP client
aiohttp==3.9.1              # Async HTTP client
httpx==0.25.2               # Modern HTTP client
responses==0.24.1           # HTTP request mocking
aioresponses==0.7.4         # Async HTTP mocking

# Data Validation and Manipulation
jsonschema==4.19.2          # JSON schema validation
cerberus==1.3.5             # Data validation
pandas==2.1.3               # Data analysis
numpy==1.25.2               # Numerical computing

# Performance and System Monitoring
psutil==5.9.6               # System resource monitoring
memory-profiler==0.61.0     # Memory usage profiling
py-spy==0.3.14              # Python profiler

# Mocking and Test Data
factory-boy==3.3.0          # Test data factories
Faker==20.1.0               # Fake data generation
freezegun==1.2.2            # Time mocking
responses-mock==1.1.1       # HTTP response mocking

# Async Testing Support
pytest-trio==0.8.0          # Trio async framework testing
pytest-tornado==0.8.1       # Tornado async testing

# Code Quality and Coverage
coverage[toml]==7.3.2       # Code coverage measurement
pytest-cov==4.1.0           # Coverage pytest plugin
coverage-badge==1.1.0       # Coverage badges

# API Testing Utilities
tavern==1.24.0              # API testing framework
locust==2.17.0              # Load testing

# Development and Debugging
ipdb==0.13.13               # Enhanced debugger
pdbpp==0.10.3               # Pretty debugger
icecream==2.1.3             # Debug printing

# Configuration and Environment
python-dotenv==1.0.0        # Environment variable management
pytest-env==1.1.0           # Environment variable testing

# Logging and Monitoring
structlog==23.2.0           # Structured logging
loguru==0.7.2               # Enhanced logging

# Time and Date Testing
freezegun==1.2.2            # Time manipulation
delorean==1.0.0             # Date/time utilities

# Network Testing
requests-mock==1.11.0       # HTTP request mocking
wiremock==2.35.0            # HTTP service mocking

# Security Testing
cryptography==41.0.7        # Cryptographic testing
pyjwt==2.8.0                # JWT token testing

# Data Generation and Fixtures
factory-boy==3.3.0          # Model factories
model-bakery==1.17.0        # Django model factories

# API Rate Limiting Testing
ratelimit==2.2.1            # Rate limiting utilities

# WebSocket Testing (for real-time data streams)
websockets==12.0            # WebSocket client
pytest-websockets==0.1.0    # WebSocket testing

# Mathematical and Financial Testing
scipy==1.11.4               # Scientific computing
statsmodels==0.14.0         # Statistical analysis
TA-Lib==0.4.28              # Technical analysis (if available)

# XML and Data Format Testing
lxml==4.9.3                 # XML processing
pytest-datafiles==3.0.0     # Test data files

# Parallel Processing Testing
joblib==1.3.2               # Parallel processing
multiprocessing-logging==0.3.4  # MP logging

# Container Health Checking
docker-compose==1.29.2      # Docker compose Python interface

# API Documentation Testing
dredd==8.2.5                # API documentation testing (if Node.js available)

# Type Checking for Tests
mypy==1.7.1                 # Static type checking
types-requests==2.31.0.10   # Type stubs for requests
types-redis==4.6.0.11       # Type stubs for redis

# Test Data Serialization
pickle5==0.0.12             # Enhanced pickle (Python < 3.8)
orjson==3.9.10              # Fast JSON serialization

# Property-based Testing
hypothesis==6.88.4          # Property-based testing

# Test Report Enhancement
allure-pytest==2.13.2       # Allure test reporting
pytest-clarity==1.0.1       # Clear assertion error messages
pytest-sugar==0.9.7         # Better test output formatting

# Load and Stress Testing
locust==2.17.0              # Load testing framework

# CI/CD Testing Support
pytest-github-actions-annotate-failures==0.2.0
pytest-json-report==1.5.0   # JSON test reports

# Development Testing Tools
pytest-watch==4.2.0         # Auto-rerun tests on file changes
pytest-testmon==2.0.12      # Only run tests affected by changes